+++ src/main/java/com/culture/BAEUNDAY/domain/chatGPT/DTO/GptResponse.java
@@ -1,4 +1,4 @@
-package com.culture.BAEUNDAY.domain.chatGPT;
+package com.culture.BAEUNDAY.domain.chatGPT.DTO;
 
 public class GptResponse{
